modification of gamescreen
