+++ app/src/main/java/ch/sdp/vibester/activity/GameSetupActivity.kt
@@ -3,19 +3,24 @@ package ch.sdp.vibester.activity
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
+import android.view.Window
 import android.widget.AdapterView
 import android.widget.ArrayAdapter
 import android.widget.Spinner
 import androidx.appcompat.app.AppCompatActivity
 import ch.sdp.vibester.R
+import ch.sdp.vibester.helper.DisplayContents
 
 class GameSetupActivity : AppCompatActivity(), AdapterView.OnItemSelectedListener {
     var text = "One"
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        requestWindowFeature(Window.FEATURE_NO_TITLE)
+        supportActionBar?.hide()
         setContentView(R.layout.activity_game_setup_screen)
 
         val spinner: Spinner = findViewById(R.id.nb_player_spinner)
+        spinner.background = DisplayContents.borderGen(this, R.color.floral_white)
         ArrayAdapter.createFromResource(
             this,
             R.array.nb_players,
