+++ app/src/main/res/layout/activity_game_setup_screen.xml
@@ -4,19 +4,21 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#FF00E0"
+    android:background="@color/floral_white"
     tools:context=".activity.GameSetupActivity">
 
     <TextView
         android:id="@+id/getPlayerNumberText"
         android:layout_width="325dp"
-        android:layout_height="79dp"
-        android:layout_marginTop="192dp"
-        android:layout_marginBottom="87dp"
-        android:fontFamily="casual"
+        android:layout_height="126dp"
+        android:layout_marginStart="41dp"
+        android:layout_marginTop="122dp"
+        android:layout_marginEnd="45dp"
+        android:layout_marginBottom="54dp"
+        android:fontFeatureSettings="smcp"
         android:text="@string/ask_nb_players"
         android:textAlignment="center"
-        android:textSize="20sp"
+        android:textSize="30dp"
         android:textStyle="bold"
         app:layout_constraintBottom_toTopOf="@+id/nb_player_spinner"
         app:layout_constraintEnd_toEndOf="parent"
@@ -27,27 +29,35 @@
         android:id="@+id/nb_player_spinner"
         android:layout_width="204dp"
         android:layout_height="48dp"
-        android:layout_marginBottom="388dp"
+        android:layout_marginStart="103dp"
+        android:layout_marginTop="54dp"
+        android:layout_marginEnd="104dp"
+        android:layout_marginBottom="89dp"
         android:entries="@array/nb_players"
-        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintBottom_toTopOf="@+id/nb_players_selected"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.497"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/getPlayerNumberText" />
 
     <Button
         android:id="@+id/nb_players_selected"
         android:layout_width="200dp"
-        android:layout_height="55dp"
-        android:layout_marginTop="81dp"
-        android:layout_marginBottom="100dp"
-        android:backgroundTint="#90953D"
+        android:layout_height="48dp"
+        android:layout_marginStart="96dp"
+        android:layout_marginTop="62dp"
+        android:layout_marginEnd="115dp"
+        android:layout_marginBottom="217dp"
+        android:backgroundTint="@color/cg_blue"
+        android:elevation="10dp"
         android:onClick="proceedToGame"
+        android:stateListAnimator="@null"
         android:text="@string/proceed"
-        app:icon="@android:drawable/btn_star_big_on"
+        android:translationZ="10dp"
+        app:icon="@android:drawable/ic_media_play"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.507"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/nb_player_spinner"
-        app:layout_constraintVertical_bias="1.0" />
+        app:layout_constraintTop_toBottomOf="@+id/nb_player_spinner" />
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
