+++ app/src/main/res/layout/activity_gamescreen.xml
@@ -3,6 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
+    android:background="@color/floral_white"
     android:layout_height="match_parent">
 
     <LinearLayout
@@ -50,6 +51,10 @@
             android:layout_width="125dp"
             android:layout_height="40dp"
             android:gravity="center"
+            android:elevation="10dp"
+            android:stateListAnimator="@null"
+            android:backgroundTint="@color/cg_blue"
+            android:translationZ="10dp"
             android:text="@string/correct" />
 
         <Button
@@ -57,6 +62,10 @@
             android:layout_width="125dp"
             android:layout_height="40dp"
             android:gravity="center"
+            android:elevation="10dp"
+            android:stateListAnimator="@null"
+            android:backgroundTint="@color/cg_blue"
+            android:translationZ="10dp"
             android:text="@string/wrong" />
 
     </LinearLayout>
@@ -80,6 +89,10 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:onClick="switchToEnding"
+        android:elevation="10dp"
+        android:stateListAnimator="@null"
+        android:backgroundTint="@color/cg_blue"
+        android:translationZ="10dp"
         android:text="@string/go_to_end"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
